# specs/001-support-crm-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Customer Support CRM API
  version: 1.0.0
  description: |
    REST contract for the Customer Support CRM (feature 001-support-crm-mvp).
    - Base path: /api/v1
    - Auth: Bearer JWT (staff & portal users) or `X-Api-Key` (integrations).
    - Errors: RFC 7807 Problem Details (`application/problem+json`) with `correlationId`
      and, for validation, an `errors` map.
    - Lists: `page` (1-based), `pageSize` (default 25, max 100) → `PagedResult`.
    - Concurrency: mutable resources return `version` (base64 rowversion); send it back on
      PUT/PATCH; stale version → 409.
    - Localization: `Accept-Language: ar|en` for messages; bilingual fields are `nameEn`/`nameAr`.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
  - apiKey: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Sign in (staff or portal customer)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { description: Signed in, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResult' } } } }
        '202': { description: Two-factor code required, content: { application/json: { schema: { $ref: '#/components/schemas/TwoFactorChallenge' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Rotate refresh token (HttpOnly cookie) and get a new access token
      responses:
        '200': { description: Refreshed, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResult' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /tickets/{ticketId}:
    parameters: [ { $ref: '#/components/parameters/TicketId' } ]
    get:
      tags: [Tickets]
      responses:
        '200': { description: Ticket, content: { application/json: { schema: { $ref: '#/components/schemas/Ticket' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Tickets]
      summary: Update subject, category, priority, department
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/TicketUpdate' } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Ticket' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
  /tickets/{ticketId}/assign:
    parameters: [ { $ref: '#/components/parameters/TicketId' } ]
    post:
      tags: [Tickets]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                assigneeId: { type: [string, 'null'], format: uuid, description: null = return to queue }
                departmentId: { type: string, format: uuid }
                version: { type: string }
      responses:
        '200': { description: Assigned, content: { application/json: { schema: { $ref: '#/components/schemas/Ticket' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /tickets/{ticketId}/take:
    parameters: [ { $ref: '#/components/parameters/TicketId' } ]
    post:
      tags: [Tickets]
      summary: Agent takes an unassigned ticket from their department queue
      responses:
        '200': { description: Taken, content: { application/json: { schema: { $ref: '#/components/schemas/Ticket' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
  /tickets/{ticketId}/status:
    parameters: [ { $ref: '#/components/parameters/TicketId' } ]
    post:
      tags: [Tickets]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status, version]
              properties:
                status: { $ref: '#/components/schemas/TicketStatus' }
                reason: { type: string, maxLength: 500 }
                version: { type: string }
      responses:
        '200': { description: Status changed, content: { application/json: { schema: { $ref: '#/components/schemas/Ticket' } } } }
        '409': { description: Invalid transition or stale version, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
  /tickets/{ticketId}/escalate:
    parameters: [ { $ref: '#/components/parameters/TicketId' } ]
    post:
      tags: [Tickets]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [reason, version]
              properties:
                reason: { type: string, minLength: 3, maxLength: 500 }
                version: { type: string }
      responses:
        '200': { description: Escalated, content: { application/json: { schema: { $ref: '#/components/schemas/Ticket' } } } }
  /tickets/{ticketId}/history:
    parameters: [ { $ref: '#/components/parameters/TicketId' } ]
    get:
      tags: [Tickets]
      responses:
        '200': { description: History, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/TicketHistoryEntry' } } } } }
  /tickets/{ticketId}/notes:
    parameters: [ { $ref: '#/components/parameters/TicketId' } ]
    post:
      tags: [Tickets]
      summary: Internal note with optional mentions (never visible to customers)
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/InternalNoteCreate' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
...
  /tickets/{ticketId}/messages:
    parameters: [ { $ref: '#/components/parameters/TicketId' } ]
    get:
      tags: [Messages]
      parameters:
        - { name: includeInternal, in: query, schema: { type: boolean, default: true } }
      responses:
        '200': { description: Conversation, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Message' } } } } }
    post:
      tags: [Messages]
      summary: Reply to customer (sent on ticket channel unless overridden)
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/ReplyCreate' } } }
      responses:
        '202': { description: Queued for delivery, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
...
  /portal/tickets/{ticketId}:
    parameters: [ { $ref: '#/components/parameters/TicketId' } ]
    get:
      tags: [Portal]
      summary: Ticket with public messages only
      responses:
        '200': { description: Ticket, content: { application/json: { schema: { $ref: '#/components/schemas/PortalTicket' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /portal/tickets/{ticketId}/messages:
    parameters: [ { $ref: '#/components/parameters/TicketId' } ]
    post:
      tags: [Portal]
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [body], properties: { body: { type: string, maxLength: 20000 }, attachmentIds: { type: array, items: { type: string, format: uuid } } } }
      responses:
        '201': { description: Reply added, content: { application/json: { schema: { $ref: '#/components/schemas/PortalMessage' } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    apiKey: { type: apiKey, in: header, name: X-Api-Key }
  parameters:
...
    TicketId: { name: ticketId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    Unauthorized:
      description: Not authenticated
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
    Forbidden:
      description: Not permitted
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
    NotFound:
      description: Not found (also returned for records outside the caller's data scope)
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
    Conflict:
      description: Stale version or invalid state; body includes `currentVersion`
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
    TooManyRequests:
      description: Rate limited
      headers: { Retry-After: { schema: { type: integer } } }
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
...
  schemas:
...
    ProblemDetails:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, description: 'Stable machine code, e.g. ticket.invalid-transition' }
        correlationId: { type: string }
        currentVersion: { type: string }
...
    Language: { type: string, enum: [ar, en] }
    Channel: { type: string, enum: [Email, WhatsApp, Sms, LiveChat, WebForm, Portal, Phone, Api] }
    UserRef:
      type: object
      properties:
        id: { type: string, format: uuid }
        fullName: { type: string }
    NamedRef:
      type: object
      properties:
        id: { type: string, format: uuid }
        nameEn: { type: string }
        nameAr: { type: string }

    # ---------- auth
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    TwoFactorChallenge:
      type: object
      properties:
        challengeId: { type: string }
    AuthResult:
      type: object
      required: [accessToken, expiresIn, user]
      properties:
        accessToken: { type: string }
        expiresIn: { type: integer, description: seconds }
        user: { $ref: '#/components/schemas/CurrentUser' }
    CurrentUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        fullName: { type: string }
        email: { type: string }
        userType: { type: string, enum: [Staff, Customer] }
        roles: { type: array, items: { type: string } }
        permissions: { type: array, items: { type: string } }
        departmentIds: { type: array, items: { type: string, format: uuid } }
        branchId: { type: [string, 'null'], format: uuid }
        preferredLanguage: { $ref: '#/components/schemas/Language' }

    # ---------- customers
...
    ContactPersonUpsert:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 150 }
        email: { type: [string, 'null'], format: email }
        phone: { type: [string, 'null'] }
        jobTitle: { type: [string, 'null'] }
        isPrimary: { type: boolean }
    ContactPerson:
      allOf:
        - $ref: '#/components/schemas/ContactPersonUpsert'
        - type: object
          properties: { id: { type: string, format: uuid } }
...
    AttachmentInfo:
      type: object
      properties:
        id: { type: string, format: uuid }
        fileName: { type: string }
        contentType: { type: string }
        sizeBytes: { type: integer, maximum: 10485760 }
        downloadUrl: { type: string }
...
    TicketStatus: { type: string, enum: [New, Open, InProgress, PendingCustomer, Resolved, Closed] }
    TicketPriority: { type: string, enum: [Low, Medium, High, Urgent] }
    SlaInfo:
      type: object
      properties:
        firstResponseDueAt: { type: [string, 'null'], format: date-time }
        resolutionDueAt: { type: [string, 'null'], format: date-time }
        firstRespondedAt: { type: [string, 'null'], format: date-time }
        isPaused: { type: boolean }
        state: { type: string, enum: [onTrack, atRisk, breached, met, none] }
        firstResponseBreached: { type: boolean }
        resolutionBreached: { type: boolean }
...
    TicketUpdate:
      type: object
      required: [version]
      properties:
        subject: { type: string, maxLength: 250 }
        categoryId: { type: string, format: uuid }
        priority: { $ref: '#/components/schemas/TicketPriority' }
        departmentId: { type: string, format: uuid }
        version: { type: string }
    TicketSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        referenceNumber: { type: string }
        subject: { type: string }
        status: { $ref: '#/components/schemas/TicketStatus' }
        priority: { $ref: '#/components/schemas/TicketPriority' }
        channel: { $ref: '#/components/schemas/Channel' }
        customer: { type: object, properties: { id: { type: string, format: uuid }, name: { type: string } } }
        assignee: { $ref: '#/components/schemas/UserRef' }
        category: { $ref: '#/components/schemas/NamedRef' }
        department: { $ref: '#/components/schemas/NamedRef' }
        sla: { $ref: '#/components/schemas/SlaInfo' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Ticket:
      allOf:
        - $ref: '#/components/schemas/TicketSummary'
        - type: object
          properties:
            description: { type: string }
            contactPerson: { $ref: '#/components/schemas/ContactPerson' }
            branch: { $ref: '#/components/schemas/NamedRef' }
            escalationLevel: { type: integer }
            resolvedAt: { type: [string, 'null'], format: date-time }
            closedAt: { type: [string, 'null'], format: date-time }
            aiSuggestion: { $ref: '#/components/schemas/AiClassification' }
            attachments: { type: array, items: { $ref: '#/components/schemas/AttachmentInfo' } }
            allowedTransitions: { type: array, items: { $ref: '#/components/schemas/TicketStatus' } }
            version: { type: string }
...
    TicketHistoryEntry:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        user: { $ref: '#/components/schemas/UserRef' }
        changeType: { type: string, enum: [Created, FieldChanged, Assigned, StatusChanged, Escalated, SlaBreached, Merged, Reopened] }
        field: { type: [string, 'null'] }
        oldValue: { type: [string, 'null'] }
        newValue: { type: [string, 'null'] }
        reason: { type: [string, 'null'] }

    # ---------- messages
    Message:
      type: object
      properties:
        id: { type: string, format: uuid }
        ticketId: { type: string, format: uuid }
        direction: { type: string, enum: [Inbound, Outbound] }
        isInternal: { type: boolean }
        isFromBot: { type: boolean }
        channel: { $ref: '#/components/schemas/Channel' }
        body: { type: string }
        sender: { type: object, properties: { kind: { type: string, enum: [Agent, Customer, System, Bot] }, id: { type: [string, 'null'], format: uuid }, name: { type: string } } }
        mentions: { type: array, items: { $ref: '#/components/schemas/UserRef' } }
        attachments: { type: array, items: { $ref: '#/components/schemas/AttachmentInfo' } }
        deliveryStatus: { type: [string, 'null'], enum: [Pending, Sent, Delivered, Read, Failed, null] }
        deliveryError: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
    ReplyCreate:
      type: object
      required: [body]
      properties:
        body: { type: string, maxLength: 20000 }
        channel: { $ref: '#/components/schemas/Channel' }
        attachmentIds: { type: array, items: { type: string, format: uuid } }
        setStatus: { $ref: '#/components/schemas/TicketStatus' }
        usedAiSuggestion: { type: boolean }
    InternalNoteCreate:
      type: object
      required: [body]
      properties:
        body: { type: string, maxLength: 20000 }
        mentionUserIds: { type: array, items: { type: string, format: uuid } }

    # ---------- dashboard / tasks / quick replies / notifications
...
    AiClassification:
      type: object
      properties:
        categoryId: { type: [string, 'null'], format: uuid }
        priority: { $ref: '#/components/schemas/TicketPriority' }
        confidence: { type: number, minimum: 0, maximum: 1 }
        rationale: { type: string }

    # ---------- portal / public
...
    PortalMessage:
      type: object
      description: Public message only (internal notes never included)
      properties:
        id: { type: string, format: uuid }
        direction: { type: string, enum: [Inbound, Outbound] }
        body: { type: string }
        senderName: { type: string }
        attachments: { type: array, items: { $ref: '#/components/schemas/AttachmentInfo' } }
        createdAt: { type: string, format: date-time }
    PortalTicket:
      type: object
      properties:
        id: { type: string, format: uuid }
        referenceNumber: { type: string }
        subject: { type: string }
        status: { $ref: '#/components/schemas/TicketStatus' }
        category: { $ref: '#/components/schemas/NamedRef' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        messages: { type: array, items: { $ref: '#/components/schemas/PortalMessage' } }
        canGiveFeedback: { type: boolean }
...